pyproject.toml
[tool.poetry]
name = "test-automation-platform"
version = "0.1.0"
description = ""
authors = ["hyunjongoh <oasis1418@gmail.com>"]
readme = "README.md"
packages = [
    { include='src', from='.' }
]

[tool.poetry.dependencies]
python = "^3.10"
Appium-Python-Client = "2.7.0"
chromedriver-autoinstaller = "^0.4.0"
pylint = "^2.15.4"
black = "^22.10.0"
pyright = "^1.1.276"
pages = "^0.3"
hvac = "^1.0.2"
requests = "^2.28.2"
slack-sdk = "^3.19.5"
webdriver-manager = "^3.8.5"
testrail-api = "^1.12.0"

[tool.poetry.scripts]
startup = "src.run:startup"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"